Generalise the test InetAddressResolverProvider (not multi-tenant specific) and use domain names that conform to RFC 2606.
