+++ integrationtests/src/test/resources/META-INF/services/java.net.spi.InetAddressResolverProvider
@@ -1,1 +1,1 @@
-io.kroxylicious.net.MultiTenantSniInetAddressResolverProvider
+io.kroxylicious.net.IntegrationTestInetAddressResolverProvider
